% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-x.R
\name{clip_to_support}
\alias{clip_to_support}
\title{Clip a vector to a support range (finite bounds only)}
\usage{
clip_to_support(x, support)
}
\description{
Clip a vector to a support range (finite bounds only)
}
\keyword{internal}
