% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-describe.R
\name{.format_panel_composition_table}
\alias{.format_panel_composition_table}
\title{Format panel composition table for printing}
\usage{
.format_panel_composition_table(x)
}
\description{
Format panel composition table for printing
}
\keyword{internal}
