% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-impute.R
\name{.get_panel_classification}
\alias{.get_panel_classification}
\title{Extract panel classification table}
\usage{
.get_panel_classification(x, id = "EIN2")
}
\description{
Extract panel classification table
}
\keyword{internal}
