% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-der.R
\name{get_der}
\alias{get_der}
\title{Debt to Equity Ratio}
\usage{
get_der(df, debt, equity, winsorize = 0.98)
}
\arguments{
\item{debt}{Total debt, EOY (On 990: Part X, line 17B; On EZ: Not Available).}

\item{equity}{Unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
The original dataframe appended with the debt to equity ratio (\code{der}),
a winsorized version (\code{der.w}), a standardized z-score version (\code{der.z}),
and a percentile version (\code{der.p}).
}
\description{
Calculate the debt to equity ratio and append it to the dataframe.
}
\details{
Debt to Equity Ratio is a a leverage ratio that defines the total amount of debt relative to the
unrestricted assets owned by an organization.This metric shows the big picture view of how much an organization
owes relative to what it owns.Higher values mean it is more highly leveraged, meaning it has low capacity for future
borrowing. As an example: if an organization has a total-debt-to-net-assets ratio of 0.4, 40\% of its assets
are financed by creditors, and 60\% are its own unrestricted, available equity. As this percentage
creeps above the 50\% mark, it can call into question the organization’s ability to manage debt,
which could jeopardize the delivery of programs and services. However, for developers, extremely
low values may mean the organization is not capitalizing enough on its equity to expand.
This value can be negative if an organization has either overpaid on its debts or if it has negative
unrestricted net assets.One limitation of this metric is that it does not provide any indication of
asset quality or liquidity since it lumps tangible and intangible assets together.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x2[ c(15,300,600) ] <- 0
dat <- data.frame( x1,x2 )
d <- get_der( df=dat, debt="x1", equity="x2" )
head( d )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_der( df=dat, debt="x1", equity="x2", winsorize=0.95 )

}
