% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-x.R
\name{rank_normal_stable}
\alias{rank_normal_stable}
\title{Rank-based inverse normal transform (mid-ranks, \code{(rank - 0.5) / n})}
\usage{
rank_normal_stable(x)
}
\description{
Rank-based inverse normal transform (mid-ranks, \code{(rank - 0.5) / n})
}
\keyword{internal}
