% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-x.R
\name{rescale_01}
\alias{rescale_01}
\title{Rescale a numeric vector to [0, 1] given explicit lo/hi}
\usage{
rescale_01(x, lo, hi)
}
\description{
Rescale a numeric vector to [0, 1] given explicit lo/hi
}
\keyword{internal}
