% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-x.R
\name{rescale_range}
\alias{rescale_range}
\title{Rescale a numeric vector to an arbitrary [a, b] range}
\usage{
rescale_range(x, to = c(1, 100))
}
\description{
Rescale a numeric vector to an arbitrary [a, b] range
}
\keyword{internal}
