% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-compat.R
\name{efile_tables}
\alias{efile_tables}
\title{List known efile table names}
\usage{
efile_tables(cardinality = "all")
}
\arguments{
\item{cardinality}{One of \code{"all"}, \code{"1x1"}, \code{"1xm"}, or
\code{"supplemental"}.}
}
\value{
Character vector of known canonical table names.
}
\description{
List known efile table names
}
