% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-aer.R
\name{get_aer}
\alias{get_aer}
\title{Administration Expense Ratio}
\usage{
get_aer(df, mgmt.ge, total.expense, winsorize = 0.98)
}
\arguments{
\item{df}{A dataframe containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{mgmt.ge}{A character string indicating the column name for management and general expenses (On 990: Part 9, Line 25(C); On EZ: Not Available.}

\item{total.expense}{A character string indicating the column name for total expenses (On 990: Part 9, Line 25(A); On EZ: Part 1, Line 17).}
}
\value{
The original dataframe appended with the administration expense ratio (\code{aer}),
a winsorized version (\code{aer.w}), a standardized z-score version (\code{aer.z}),
and a percentile version (\code{aer.p}).
}
\description{
Calculate the administration expense ratio and append it to the dataframe.
}
\details{
The administrative expense ratio measures the percentage of an organization’s expenses that are
being allocated to administrative costs. High values in this ratio indicate that more of an organization's
expenses are going towards its management and general expenses (or overhead) while a lower number indicates
that an organization's funds are going towards program service or fundraising expenses. Charity Navigator
generally gives its highest rankings to organizations that spend less than 15\% of expenses on overhead. The
Better Business Bureau’s Wise Giving Alliance recommends a ratio of less than 35\%.
}
\examples{
x1<-rnorm( 1000,100,30 )
x2<-rnorm( 1000,200,30 )
x2[ c(1,10,100) ] <- 0

dat<-data.frame( x1, x2 )

a<-get_aer( df=dat, mgmt.ge='x1', total.expense='x2' )

head( a )

}
