% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-arr.R
\name{get_arr}
\alias{get_arr}
\title{Assets to Revenues Ratio}
\usage{
get_arr(df, total.assets, total.revenue, winsorize = 0.98)
}
\arguments{
\item{df}{A dataframe containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{total.assets}{A character string indicating the column name for total assets (On 990: Part X, Line 16B; On EZ: Pt II, Line 25B).}

\item{total.revenue}{A character string indicating the column name for total revenue (On 990: (Part VIII, Line 12A); On EZ: Part I, Line 9).}
}
\value{
The original dataframe appended with the assets to revenues ratio (\code{arr}),
a winsorized version (\code{arr.w}), a standardized z-score version (\code{arr.z}),
and a percentile version (\code{arr.p}).
}
\description{
Calculate the assets to revenues ratio and append it to the dataframe.
}
\details{
This metric measures how many dollars of assets it has per dollar of annual revenues (how many
dollars of revenue have been converted into assets in the past or that year for the current year). Generally,
this ratio should be higher for organizations with large inventories of developments or large endowments.
}
\examples{
x1<-rnorm( 1000,100,30 )
x2<-rnorm( 1000,200,30 )
x2[ c(1,10,100) ] <- 0

dat<-data.frame( x1, x2 )

a<-get_arr( df=dat, total.assets='x1', total.revenue='x2' )

head( a )

}
