% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-arr.R
\name{get_arr}
\alias{get_arr}
\title{Assets to Revenues Ratio}
\usage{
get_arr( df, total.assets = "F9_10_ASSET_TOT_EOY", 
total.revenue = c( "F9_08_REV_TOT_TOT", "F9_01_REV_TOT_CY" ), winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{total.assets}{A character string indicating the column name for total assets (On 990: Part X, Line 16B; On EZ: Pt II, Line 25B).}

\item{total.revenue}{A character string indicating the column name for total revenue (On 990: (Part VIII, Line 12A); On EZ: Part I, Line 9).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the assets to revenues ratio (\code{arr}),
a winsorized version (\code{arr.w}), a stanarrdized z-score version (\code{arr.z}),
and a percentile version (\code{arr.p}).
}
\description{
Calculate the assets to revenues ratio and append it to the dataframe.
}
\details{
This metric measures how many dollars of assets it has per dollar of annual revenues (how many
dollars of revenue have been converted into assets in the past or that year for the current year). Generally,
this ratio should be higher for organizations with large inventories of developments or large endowments.
Note: computation of this metric is available to both 990 and 990-EZ filers.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,100,30 )

dat <- data.frame( x1, x2 )

# specify own column names
d <- get_arr( df = dat, total.assets = "x1", total.revenue = "x2" )


head( d )

# run with default column names
x3 <- rnorm( 1000,100,30 )
x3[ seq( from = 1, to = 1000, 50 ) ] <- NA

dat_01 <- data.frame( x1, x2, x3 )

colnames( dat_01 ) <- c( "F9_10_ASSET_TOT_EOY", "F9_08_REV_TOT_TOT",
                         "F9_01_REV_TOT_CY" )

d <- get_arr( dat_01 )

# specify only PC variables
d <- get_arr( dat_01, total.assets = "F9_10_ASSET_TOT_EOY", total.revenue = "F9_08_REV_TOT_TOT" )

# coerce one column to factor
dat_01$F9_10_ASSET_TOT_EOY <- as.factor( dat_01$F9_10_ASSET_TOT_EOY )

d <- get_arr( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_arr( df = dat, total.assets = "x1", total.revenue ="x2", winsorize=0.95 )

d <- get_arr( dat_01, winsorize = 0.95 )

# assume only one PC variable for the denominator is present in the dataset and we run with default parameters
dat_02 <- dat_01

colnames( dat_02 ) <- c( "F9_10_ASSET_TOT_EOY", "F9_08_REV_TOT_TOT",
                         "x" )

d <- get_arr( dat_02, winsorize = 0.95 )

colnames( dat_02 ) <- c( "F9_10_ASSET_TOT_EOY", "x",
                         "F9_01_REV_TOT_CY" )

d <- get_arr( dat_02, winsorize = 0.95 )

# assume denominator variables are default names but numerator is defined
colnames( dat_02 ) <- c( "x1", "F9_08_REV_TOT_TOT",
                         "F9_01_REV_TOT_CY" )

d <- get_arr( dat_02, total.assets = "x1", winsorize = 0.95 )


# using 990 data
data( part010810 )
d <- get_arr(df=part010810)

# now coerce one of the variables to numeric
part010810$F9_10_ASSET_TOT_EOY <- as.character( part010810$F9_10_ASSET_TOT_EOY )

d <- get_arr(df=part010810)

\dontrun{
## Errors ##

# numerator not specified
d <- get_arr( df = dat, total.assets = NULL, total.revenue = "x2")

# denominator not specified
d <- get_arr( df = dat, total.assets = "x1", total.revenue = NULL )

# neither numerator nor denominator specified
d <- get_arr( df = dat, total.assets = NULL, total.revenue = NULL )
}

}
