% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-cr.R
\name{get_cr}
\alias{get_cr}
\title{Current Ratio}
\usage{
get_cr(df, assets, liabilities, winsorize = 0.98)
}
\arguments{
\item{assets}{Current Assets (On 990: (Part X, line 1B) + (Part X, line 2B) + (Part X, line 3B) + (Part X, line 4B) + (Part X, line 8B) + (Part X, line 9B); On EZ: Part I, line 22 (cash and short-term investments only).}

\item{liabilities}{Current Liabilities (On 990: (Part X, line 17B) + (Part X, line 18B); On EZ: Not available).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
The original dataframe with the current ratio (\code{cr}),
a winsorized version (\code{cr.w}), a standardized z-score version (\code{cr.z}),
and a percentile version (\code{cr.p}).
}
\description{
Calculate the current ratio and append it to the dataframe.
}
\details{
The current ratio is used to measure the overall liquidity of a nonprofit organization.
In its simplest form, it shows how many dollars of current assets an organization has to cover its
current obligations. The higher the ratio, the more liquid the organization.
As a rule of thumb, organizations should strive for a current ratio of 1.0 or higher. An organization
with a ratio of 1.0 would have one dollar of assets to pay for every dollar of current liabilities.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x2[ c(15,300,600) ] <- 0
dat<-data.frame( x1,x2 )
d <- get_ssr( df=dat, assets='x1', liabilities='x2' )
head( d )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_ssr( df=dat, assets='x1', liabilities='x2', winsorize=0.95 )
  @export
}
