% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package.R
\name{get_dar}
\alias{get_dar}
\title{Debt to Asset Ratio}
\usage{
get_dar(df, debt, assets, winsorize = 0.98)
}
\arguments{
\item{debt}{Total liabilities (On 990: Part X, line 26B; On EZ: Part II, line 26B).}

\item{assets}{Total assets, EOY (On 990: Part X, line 16B; On EZ: Part II, line 25B).}

\item{winsorize}{The winsorization value,}
}
\value{
The original dataframe appended with the debt to asset ratio (\code{dar}),
a winsorized version (\code{dar.w}), a standardized z-score version (\code{dar.z}),
and a percentile version (\code{dar.p}).
}
\description{
Calculate the debt to asset ratio and append it to the dataframe.
}
\details{
Total-debt-to-total-assets is a leverage ratio that defines the total
amount of debt relative to assets owned by a company. Using this metric,
analysts can compare one company's leverage with that of other companies
in the same industry. This information can reflect how financially stable
a company is. The higher the ratio, the higher the degree of leverage (DoL) and,
consequently, the higher the risk (\href{Investopedia}{https://www.investopedia.com/terms/t/totaldebttototalassets.asp}).
}
\examples{
x1 <- rnorm(1000,100,30)
x2 <- rnorm(1000,200,30)
x2[ c(15,300,600) ] <- 0
dat <- data.frame(x1,x2)
d <- get_dar( df=dat, debt="x1", assets="x2" )
head( d )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_dar( df=dat, debt="x1", assets="x2", winsorize=0.95 )

}
