% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-der.R
\name{get_der}
\alias{get_der}
\title{Debt to Equity Ratio}
\usage{
get_der( df, debt = "F9_10_LIAB_ACC_PAYABLE_EOY", equity = "F9_10_NAFB_UNRESTRICT_EOY", winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{debt}{A character string indicating the column name for total debt, EOY (On 990: Part X, line 17B; On EZ: Not Available) with the default name supplied.}

\item{equity}{A character string indicating the column name for unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available) with the default name supplied.}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the debt to equity ratio (\code{der}),
a winsorized version (\code{der.w}), a standardized z-score version (\code{der.z}),
and a percentile version (\code{der.p}).
}
\description{
Calculate the debt to equity ratio and append it to the dataframe.
}
\details{
Debt to Equity Ratio is a a leverage ratio that defines the total amount of debt relative to the
unrestricted assets owned by an organization.This metric shows the big picture view of how much an organization
owes relative to what it owns.Higher values mean it is more highly leveraged, meaning it has low capacity for future
borrowing. As an example: if an organization has a total-debt-to-net-assets ratio of 0.4, 40\% of its assets
are financed by creditors, and 60\% are its own unrestricted, available equity. As this percentage
creeps above the 50\% mark, it can call into question the organization’s ability to manage debt,
which could jeopardize the delivery of programs and services. However, for developers, extremely
low values may mean the organization is not capitalizing enough on its equity to expand.
This value can be negative if an organization has either overpaid on its debts or if it has negative
unrestricted net assets.One limitation of this metric is that it does not provide any indication of
asset quality or liquidity since it lumps tangible and intangible assets together. Note: computation of this metric
is available to only 990 filers and not for 990-EZ filers. The default inputs use column names for variables
available only to 990 filers. Note: This ratio can be interchanged with total liabilities over total net assets
(which should be comparable for EZ filers and full 990 filers), but for Community Development Corporations, the
more important metric is unrestricted net assets, which isn’t available for EZ filers.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000, 100, 30 )
x2 <- rnorm( 1000, 200, 30 )
x2[ c( 15, 300, 600 ) ] <- 0

dat <- data.frame( x1, x2 )

# specify own column names
d <- get_der( df = dat, debt = "x1", equity = "x2" )


head( d )

# run with default column names
x1[ seq( from = 1, to = 1000, 50 ) ] <- NA
x2[ seq( from = 1, to = 1000, 71 ) ] <- NA

dat_01 <- data.frame( x1, x2 )

colnames( dat_01 ) <- c( "F9_10_LIAB_ACC_PAYABLE_EOY", "F9_10_NAFB_UNRESTRICT_EOY" )

d <- get_der( dat_01 )

# coerce one column to factor
dat_01$F9_10_LIAB_ACC_PAYABLE_EOY <- as.factor( dat_01$F9_10_LIAB_ACC_PAYABLE_EOY )

d <- get_der( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_der( df = dat, debt = "x1", equity = "x2", winsorize = 0.95 )

d <- get_der( dat_01, winsorize = 0.95 )


# using 990 data
data( part010810 )
d <- get_der( df = part010810 )

# now coerce one of the variables to numeric
part010810$F9_01_EXP_TOT_CY <- as.character( part010810$F9_01_EXP_TOT_CY )

d <- get_der( df = part010810 )

\dontrun{
## Errors ##

# numerator not specified
d <- get_der( df = dat, debt = NULL, equity = "x2" )

# denominator not specified
d <- get_der( df = dat, debt = "x1", equity = NULL )

# neither numerator nor denominator specified
d <- get_der( df = dat, debt = NULL, equity = NULL )

# column names vector not of correct length
d <- get_der( df = dat, debt = c("e","b","c"), equity = "e")

# column names vector not of correct length
d <- get_der( df = dat, debt = "e", equity = c( "e", "b", "c") )
}
}
