% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-dgdr.R
\name{get_dgdr}
\alias{get_dgdr}
\title{Donation/Grant Dependence Ratio}
\usage{
get_dgdr(df, total.contributions, fund.income, total.revenue, winsorize = 0.98)
}
\arguments{
\item{df}{A dataframe containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{total.contributions}{A character string indicating the column name for total contributions (On 990: Part VIII, Line 1h(A); On EZ: Not Available).}

\item{fund.income}{A character string indicating the column name for fundraising income (On 990: Part VIII, Line 8c(A); On EZ: Not Available).}

\item{total.revenue}{A character string indicating the column name for total revenue (On 990: Part VIII, Line 12A; On EZ: Part I, Line 9).}
}
\value{
The original dataframe appended with the donation/grant dependence ratio (\code{dgdr}),
a winsorized version (\code{dgdr.w}), a standardized z-score version (\code{dgdr.z}),
and a percentile version (\code{dgdr.p}).
}
\description{
Calculate the donation/grant dependence ratio and append it to the dataframe.
}
\details{
This metric shows how much of an organization’s total revenues come from contributions, government
grants, and special event reveues. High levels in this indicator are undesirable since that means that an
organization’s revenues are volatile insofar as it is dependent on contributions that are highly likely to
contract during economic downturns. Low values in this indicator mean an organization is not dependent on
donations.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x3 <- rnorm( 1000,200,30 )
x3[ c(1,10,100)] <- 0

dat <- data.frame( x1, x2, x3 )

a <- get_dgdr( df=dat, total.contributions='x1', fund.income='x2', total.revenue='x3', winsorize=0.98 )

head( a )

}
