% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-dmr.R
\name{get_dmr}
\alias{get_dmr}
\title{Debt Management Ratio}
\usage{
get_dmr(df, liabilities, net.assets, winsorize = 0.98)
}
\arguments{
\item{liabilities}{A character string indicating the column name for total liabilities, EOY (On 990: Part X, line 26B; On EZ: Part II, line 26B).}

\item{net.assets}{A character string indicating the column name for unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
The original dataframe appended with the debt management ratio (\code{dmr}),
a winsorized version (\code{dmr.w}), a standardized z-score version (\code{dmr.z}),
and a percentile version (\code{dmr.p}).
}
\description{
Calculate the debt management ratio and append it to the dataframe.
}
\details{
This metric measures how much an organization is relying on funding from lending entities and the
amount of free, unrestricted cash it has to pay back those loans. A high value in this metric could mean
an organization is highly leveraged and thus that it has reduced ability to borrow more money in the future.
Conversely, a low value indicates the organization may not be leveraging its assets to achieve the most
growth and impact it could or that there is a dearth of credit available from lending entities.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x2[ c(15,300,600) ] <- 0
dat <- data.frame( x1,x2 )
d <- get_dmr( df=dat, debt='x1', net.assets='x2' )
head( d )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_dmr( df=dat, debt='x1', net.assets='x2', winsorize=0.95 )

}
