% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dmr.R
\name{get_dmr}
\alias{get_dmr}
\title{Debt Management Ratio}
\usage{
get_dmr( df, liabilities = "F9_10_LIAB_TOT_EOY", 
net.assets = "F9_10_NAFB_UNRESTRICT_EOY", 
winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{liabilities}{A character string indicating the column name for total liabilities, EOY (On 990: Part X, line 26B; On EZ: Part II, line 26B) with the default name supplied.}

\item{net.assets}{A character string indicating the column name for unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available) with the default name supplied.}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the debt management ratio (\code{dmr}),
a winsorized version (\code{dmr.w}), a standardized z-score version (\code{dmr.z}),
and a percentile version (\code{dmr.p}).
}
\description{
Calculate the debt management ratio and append it to the dataframe.
}
\details{
This metric measures how much an organization is relying on funding from lending entities and the
amount of free, unrestricted cash it has to pay back those loans. A high value in this metric could mean
an organization is highly leveraged and thus that it has reduced ability to borrow more money in the future.
Conversely, a low value indicates the organization may not be leveraging its assets to achieve the most
growth and impact it could or that there is a dearth of credit available from lending entities.
Note: computation of this metric is available to both 990 and 990-EZ filers.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000, 100, 30 )
x2 <- rnorm( 1000, 200, 30 )
x2[ c( 15, 300, 600 ) ] <- 0

dat <- data.frame( x1, x2 )

# specify own column names
d <- get_dmr( df = dat, liabilities = "x1", net.assets = "x2" )


head( d )

# run with default column names
x1[ seq( from = 1, to = 1000, 50 ) ] <- NA
x2[ seq( from = 1, to = 1000, 71 ) ] <- NA

dat_01 <- data.frame( x1, x2 )

colnames( dat_01 ) <- c( "F9_10_LIAB_TOT_EOY", "F9_10_NAFB_UNRESTRICT_EOY" )

d <- get_dmr( dat_01 )

# coerce one column to factor
dat_01$F9_10_LIAB_TOT_EOY <- as.factor( dat_01$F9_10_LIAB_TOT_EOY )

d <- get_dmr( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_dmr( df = dat, liabilities = "x1", net.assets = "x2", winsorize = 0.95 )

d <- get_dmr( dat_01, winsorize = 0.95 )


# using 990 data
data( part010810 )
d <- get_dmr( df = part010810 )

# now coerce one of the variables to numeric
part010810$F9_10_LIAB_TOT_EOY <- as.character( part010810$F9_10_LIAB_TOT_EOY )

d <- get_dmr( df = part010810 )

\dontrun{
## Errors ##

# numerator not specified
d <- get_dmr( df = dat, liabilities = NULL, net.assets = "x2" )

# denominator not specified
d <- get_dmr( df = dat, liabilities = "x1", net.assets = NULL )

# neither numerator nor denominator specified
d <- get_dmr( df = dat, liabilities = NULL, net.assets = NULL )

# column names vector not of correct length
d <- get_dmr( df = dat, liabilities = c("e","b","c"), net.assets = "e")

# column names vector not of correct length
d <- get_dmr( df = dat, liabilities = "e", net.assets = c( "e", "b", "c") )
}
}
