% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-doch.R
\name{get_doch}
\alias{get_doch}
\title{Days of Operating Cash on Hand}
\usage{
get_doch(
  df,
  f.cash = NULL,
  f.si = NULL,
  f.pr = NULL,
  f.ar = NULL,
  f.tfe = NULL,
  f.dda = NULL,
  ez.csi = NULL,
  ez.toe = NULL,
  winsorize = 0.98
)
}
\arguments{
\item{f.cash}{Cash, EOY (On 990: Part X, line 1B; On EZ:Part I, line 22 (cash and short-term investments only)).}

\item{f.si}{Short-term investments, EOY (On 990: Part X, line 2B; On EZ: Not Available).}

\item{f.pr}{Pledges and grant receivables, EOY On 990: Part X, line 3B; On EZ: Not Available).}

\item{f.ar}{Accounts receivables, EOY (On 990: Part X, line 4B; On EZ: Not Available).}

\item{f.tfe}{Total functional expenses (On 990: Part IX, line 25A; On EZ: Not Available).}

\item{f.dda}{Depreciation, depletion, and amortization (On 990: Part IX, line 22A; On EZ: Not Available).}

\item{ez.csi}{Cash, savings, and investment, EOY On EZ:Part II, line 22B (cash and short-term investments only)).}

\item{ez.toe}{Total operating expenses, EOY (On EZ: Part I, line 17 (operating expenses only)).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
The original dataframe appended with the days of operating cash on hand (\code{doch}),
a winsorized version (\code{doch.w}), a standardized z-score version (\code{doch.z}),
and a percentile version (\code{doch.p}).
}
\description{
Calculate the days of operating cash on hand and append it to the dataframe.
}
\details{
Days cash on hand is the number of days that an organization can continue to pay
its operating expenses, given the amount of cash available. This metric is good to review at startup
of organization, during periods of low revenues, and prior to undertaking new major activity. The metric
is limited though in the following ways. First, it is based on an average daily cash outflow, which is
not really the case. Instead, cash tends to be spent in a lumpy manner, such as when rent or payroll
are paid. Also, management tends to take drastic action to reduce expenses as cash reserves decline,
so that the actual days of operation tend to be longer than indicated by this ratio. Thus, it is better
to use a detailed cash flow analysis to determine the precise duration of the available cash, with regular
updates.
}
