% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-doci.R
\name{get_doci}
\alias{get_doci}
\title{Days of Operating Cash and Investments}
\usage{
get_doci(df, equity, invest, lbe, mnp, daily.av.exp, dda, winsorize = 0.98)
}
\arguments{
\item{df}{A dataframe containing the required field fdoci computing the metric. The metric will be appended to this dataset.}

\item{equity}{A character string indicating the column name for unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available).}

\item{invest}{A character string indicating the column name for investment income (On 990: Part X, Line 8(A); On EZ: Not Available).}

\item{lbe}{A character string indicating the column name for lands, buildings, and equipment (On 990: Part X, Line 10(c)(B); On EZ: Not Available).}

\item{mnp}{A character string indicating the column name for mortgages and notes payables (On 990: Part X, Line 23(B); On EZ: Not Available).}

\item{daily.av.exp}{A character string indicating the column name for daily average expenses or total expenses (On 990: Part IX, Line 25(A); On EZ: Not Available).}

\item{dda}{Depreciation, depletion, and amortization (On 990: Part IX, line 22A; On EZ: Not Available).}
}
\value{
The original dataframe appended with the days of operating cash and investments (\code{doci}),
a winsorized version (\code{doci.w}), a standardized z-score version (\code{doci.z}),
and a percentile version (\code{doci.p}).
}
\description{
Calculate the days of operating cash and investments and append it to the dataframe.
}
\details{
The days of operating cash and investment measures how long an organization would be able to
continue operating its programs and investing at its current rate based on its average per-day expenses
(excluding depreciation). Note: This ratio is the same as month of cash on hand but it is in days and
excludes land and mortgages from the operating cash and investment numerator. The reason that land and
mortgage expenses are deducted from the numerator is because those are non-liquid or necessary expenses
and shouldn't truly be considered free cash for operating and investment. The higher the number of days
from this metric, the more protected the system will be against revenue shocks, but the target value is
subjective. Generally, a system should aim to maintain several months’ worth of cash on hand and at the
very least exceed the length of the billing period (usually 30 or 60 days).
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x3 <- rnorm( 1000,200,30 )
x4 <- rnorm( 1000,200,30 )
x5 <- rnorm( 1000,200,30 )
x6 <- rnorm( 1000,200,30 )

dat <- data.frame( x1, x2, x3, x4, x5, x6 )

a <- get_doci( df=dat, equity='x1', invest = 'x2',
              lbe='x3', mnp='x4', daily.av.exp='x5', dda='x6' )

# #zero in the denominator
x5[ c(1:10) ] <- 0
x6[ c(1:10) ] <- 0
dat <- data.frame( x1, x2, x3, x4, x5, x6 )

b <- get_doci( df=dat, equity='x1', invest = 'x2',
              lbe='x3', mnp='x4', daily.av.exp='x5', dda='x6' )

c <- get_doci( df=dat, equity='x1', invest = 'x2',
              lbe='x3', mnp='x4', daily.av.exp='x5', dda='x6', winsorize = 0.975 )


}
