% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-doci.R
\name{get_doci}
\alias{get_doci}
\title{Days of Operating Cash and Investments}
\usage{
get_doci( df, 
net.assets = "F9_10_NAFB_UNRESTRICT_EOY", 
invest = "F9_10_ASSET_INV_SALE_EOY", 
lbe = "F9_10_ASSET_LAND_BLDG_NET_EOY", 
mnp = "F9_10_LIAB_MTG_NOTE_EOY", 
tot.func.exp = "F9_09_EXP_TOT_TOT", 
dda = "F9_09_EXP_DEPREC_TOT", numerator = NULL, denominator = NULL, winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{net.assets}{A character string indicating the column name for unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available).}

\item{invest}{A character string indicating the column name for investment income (On 990: Part X, Line 8(B); On EZ: Not Available).}

\item{lbe}{A character string indicating the column name for lands, buildings, and equipment (On 990: Part X, Line 10(c)(B); On EZ: Not Available).}

\item{mnp}{A character string indicating the column name for mortgages and notes payables (On 990: Part X, Line 23(B); On EZ: Not Available).}

\item{tot.func.exp}{A character string indicating the column name for daily average expenses or total expenses (On 990: Part IX, Line 25(A); On EZ: Not Available).}

\item{dda}{Depreciation, depletion, and amortization (On 990: Part IX, line 22A; On EZ: Not Available).}

\item{numerator}{A character string indicating the user-supplied column name for a pre-aggregated variable for the numerator. Do not combine with numerator column component arguments (\code{net.assets}, \code{invest},\code{lbe}, \code{mnp}).}

\item{denominator}{A character string indicating the user-supplied column name for a pre-aggregated variable for the denominator. Do not combine with denominator column component arguments (\code{tot.func.exp}, \code{dda}).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the days of operating cash and investments (\code{doci}),
a winsorized version (\code{doci.w}), a standardized z-score version (\code{doci.z}),
and a percentile version (\code{doci.p}).
}
\description{
Calculate the days of operating cash and investments and append it to the dataframe.
}
\details{
The days of operating cash and investment measures how long an organization would be able to
continue operating its programs and investing at its current rate based on its average per-day expenses
(excluding depreciation). Note: This ratio is the same as month of cash on hand but it is in days and
excludes land and mortgages from the operating cash and investment numerator. The reason that land and
mortgage expenses are deducted from the numerator is because those are non-liquid or necessary expenses
and shouldn't truly be considered free cash for operating and investment. The higher the number of days
from this metric, the more protected the system will be against revenue shocks, but the target value is
subjective. Generally, a system should aim to maintain several months’ worth of cash on hand and at the
very least exceed the length of the billing period (usually 30 or 60 days). Note: computation of this metric
is available to only 990 filers and not for 990-EZ filers. The default inputs use column names for variables
available only to 990 filers.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x3 <- rnorm( 1000,200,30 )
x4 <- rnorm( 1000,100,30 )
x5 <- rnorm( 1000,200,30 )
x6 <- rnorm( 1000,200,30 )
dat <- data.frame( x1, x2, x3, x4, x5, x6 )

# specify own column names
d <- get_doci( df=dat, net.assets='x1', invest='x2', lbe='x3', mnp='x4', tot.func.exp='x5', dda = 'x6', winsorize=0.98 )

head( d )

# run with default column names
dat_01 <- dat

colnames( dat_01 ) <- c( "F9_10_NAFB_UNRESTRICT_EOY", "F9_10_ASSET_INV_SALE_EOY", "F9_10_ASSET_LAND_BLDG_NET_EOY",
                         "F9_10_LIAB_MTG_NOTE_EOY", "F9_09_EXP_TOT_TOT", "F9_09_EXP_DEPREC_TOT" )


d <- get_doci( dat_01 )

# coerce one column to factor
dat_01$F9_10_NAFB_UNRESTRICT_EOY <- as.factor( dat_01$F9_10_NAFB_UNRESTRICT_EOY )

d <- get_doci( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_doci( df=dat, net.assets='x1', invest='x2', lbe='x3', mnp='x4', tot.func.exp='x5', dda = 'x6', winsorize = 0.95 )

d <- get_doci( dat_01, winsorize = 0.95 )

# aggregate variables into single numerator and denominator
x.den <- x5 + x6
x.num <- x1 + x2 + x3 + x4

dat_02 <- cbind( dat, x.den, x.num)

d <- get_doci( dat_02, numerator = "x.num", denominator = "x.den" )

# using 990 data
data( part010810 )
d <- get_doci( df = part010810 )

# now coerce one of the variables to numeric
part010810$F9_10_NAFB_UNRESTRICT_EOY <- as.character( part010810$F9_10_NAFB_UNRESTRICT_EOY )

d <- get_doci( df = part010810 )

\dontrun{
#errors 

# incorrectly specify denominator
get_doci( df = dat_02, net.assets = 'x1', invest = 'x2', lbe = 'x3', mnp = 'x4', 
          tot.func.exp = NULL, numerator = NULL, denominator = NULL, winsorize=0.98 )

# incorrectly specify numerator with conflicting arguments
get_doci( df = dat_02, net.assets = 'x1', invest = 'x2', lbe = 'x3', mnp = 'x4', 
          tot.func.exp = NULL, dda = NULL, numerator = 'x.num', denominator = 'x.den' , winsorize=0.98 )


# incorrectly specify denominator with conflicting arguments
get_doci( df = dat_02, net.assets = 'x1', invest = 'x2', lbe = 'x3', mnp = 'x4', 
          tot.func.exp = 'x5', dda = 'x6', numerator = NULL, denominator = 'x.den' , winsorize=0.98 )  

# supplying no arguments for the numerator
get_doci( df = dat_02, net.assets = NULL, invest = NULL, lbe = NULL, mnp = NULL, 
          tot.func.exp = 'x5', dda = 'x6', numerator = NULL, denominator = NULL , winsorize=0.98 )  

# supplying no arguments for the denominator
get_doci( df = dat_02, net.assets = 'x1', invest = 'x2', lbe = 'x3', mnp = 'x4', 
          tot.func.exp = NULL, dda = NULL, numerator = NULL, denominator = NULL , winsorize=0.98 )  

# supplying argument for one of the parameters in the numerator that is greater than length 1
get_doci( df = dat_02, net.assets = 'x1', invest = 'x2', lbe = 'x3', mnp = 'x4', 
          tot.func.exp = c( 'x5', 'x6' ), dda = 'x6', numerator = NULL, denominator = NULL, winsorize=0.98 )
}
}
