% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-eidr.R
\name{get_eidr}
\alias{get_eidr}
\title{Earned Income Dependence Ratio}
\usage{
get_eidr(
  df,
  prog.service.rev,
  memb.dues,
  royalties,
  other.revenue,
  total.revenue,
  winsorize = 0.98
)
}
\arguments{
\item{df}{A dataframe containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{prog.service.rev}{A character string indicating the column name for program service revenue (On 990: Part VIII, Line 2g(A); On EZ: Not Available).}

\item{memb.dues}{A character string indicating the column name for membership dues (On 990: Part VIII, Line 1b(A); On EZ: Not Available).}

\item{royalties}{A character string indicating the column name for royalties (On 990: Part VIII, Line 5(A); On EZ: Not Available).}

\item{other.revenue}{A character string indicating the column name for all other revenue (On 990: Part VIII, Line 11d(A); On EZ: Not Available).}

\item{total.revenue}{A character string indicating the column name for total revenue (On 990: (Part VIII, Line 12A); On EZ: Part I, Line 9).}
}
\value{
The original dataframe appended with the earned income dependence ratio (\code{eidr}),
a winsorized version (\code{eidr.w}), a standardized z-score version (\code{eidr.z}),
and a percentile version (\code{eidr.p}).
}
\description{
Calculate the earned income dependence ratio and append it to the dataframe.
}
\details{
This metric shows how much of an organization’s total revenues come from earned income (program
service revenues, dues, assessments, profits from sales). High levels in this indicator are more desirable
since that means that an organization is fairly self-sustaining with its own activities. Low values in this
indicator mean an organization is likely dependent on donations or investments for their revenues and thus more
vulnerable to the sentiments of donors or market forces.
}
\examples{
x1<-rnorm( 1000,100,30 )
x2<-rnorm( 1000,200,30 )
x3<-rnorm( 1000,200,30 )
x4<-rnorm( 1000,100,30 )
x5<-rnorm( 1000,200,30 )
x5[ c(1,10,100) ] <- 0

dat<-data.frame( x1, x2, x3, x4, x5 )

a<-get_eidr( df=dat, prog.service.rev='x1', memb.dues='x2', royalties='x3', other.revenue='x4', total.revenue='x5', winsorize=0.98)

head( a )

}
