% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-er.R
\name{get_er}
\alias{get_er}
\title{Equity Ratio}
\usage{
get_er( df, net.assets = 'F9_10_NAFB_TOT_EOY', total.assets = 'F9_10_ASSET_TOT_EOY', winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset with the default name supplied.}

\item{net.assets}{A character string indicating the column name for net assets, EOY (On 990: Part X, Line 33B; On EZ: Part I, Line 21) with the default name supplied.}

\item{total.assets}{A character string indicating the column name for total assets, EOY (On 990: Part X, Line 16B; On EZ: Part II, Line 25B) with the default name supplied.}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the equity ratio (\code{er}),
a winsorized version (\code{er.w}), a standardized z-score version (\code{er.z}),
and a percentile version (\code{er.p}).
}
\description{
Calculate the equity ratio and append it to the dataframe.
}
\details{
This metric indicates how much of an organization’s assets are owned free and clear or how much
equity it has in its total assets. Nonprofits with greater amounts of equity are more flexible in the face
of financial shocks than organizations with comparatively lesser amounts of equity because they can (1)
borrow money from capital markets and (2) convert those assets to cash to offset financial shocks. High
values in this indicator are generally better, as they show that an organization has substantial equity in
its assets. Low or negative values indicate an organization has higher liabilities and is generally more
leveraged and thus more vulnerable to shocks. However, low values also indicate that an organization may
be investing more of its equity for growth. Note: computation of this metric is available to both 990 and 990-EZ filers.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000, 100, 30 )
x2 <- rnorm( 1000, 200, 30 )
x2[ c( 15, 300, 600 ) ] <- 0

dat <- data.frame( x1, x2 )

# specify own column names
d <- get_er( df = dat, net.assets = "x1", total.assets = "x2" )


head( d )

# run with default column names
x1[ seq( from = 1, to = 1000, 50 ) ] <- NA
x2[ seq( from = 1, to = 1000, 71 ) ] <- NA

dat_01 <- data.frame( x1, x2 )

colnames( dat_01 ) <- c( "F9_10_NAFB_TOT_EOY", "F9_10_ASSET_TOT_EOY" )

d <- get_er( dat_01 )

# coerce one column to factor
dat_01$F9_10_NAFB_TOT_EOY <- as.factor( dat_01$F9_10_NAFB_TOT_EOY )

d <- get_er( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_er( df = dat, net.assets = "x1", total.assets = "x2", winsorize = 0.95 )

d <- get_er( dat_01, winsorize = 0.95 )


# using 990 data
data( part010810 )
d <- get_er( df = part010810 )

# now coerce one of the variables to numeric
part010810$F9_01_EXP_TOT_CY <- as.character( part010810$F9_01_EXP_TOT_CY )

d <- get_er( df = part010810 )

\dontrun{
## Errors ##

# numerator not specified
d <- get_er( df = dat, net.assets = NULL, total.assets = "x2" )

# denominator not specified
d <- get_er( df = dat, net.assets = "x1", total.assets = NULL )

# neither numerator nor denominator specified
d <- get_er( df = dat, net.assets = NULL, total.assets = NULL )

# column names vector not of correct length
d <- get_er( df = dat, net.assets = c("e","b","c"), total.assets = "e")

# column names vector not of correct length
d <- get_er( df = dat, net.assets = "e", total.assets = c( "e", "b", "c") )
}
}
