% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-er.R
\name{get_er}
\alias{get_er}
\title{Equity Ratio}
\usage{
get_er(df, net.assets, total.assets, winsorize = 0.98)
}
\arguments{
\item{net.assets}{A character string indicating the column name for net assets, EOY (On 990: Part X, Line 33B; On EZ: Part I, Line 21).}

\item{total.assets}{A character string indicating the column name for total assets, EOY (On 990: Part X, Line 16B; On EZ: Part II, Line 25B).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
The original dataframe appended with the equity ratio (\code{er}),
a winsorized version (\code{er.w}), a standardized z-score version (\code{er.z}),
and a percentile version (\code{er.p}).
}
\description{
Calculate the equity ratio and append it to the dataframe.
}
\details{
This metric indicates how much of an organization’s assets are owned free and clear or how much
equity it has in its total assets. Nonprofits with greater amounts of equity are more flexible in the face
of financial shocks than organizations with comparatively lesser amounts of equity because they can (1)
borrow money from capital markets and (2) convert those assets to cash to offset financial shocks. High
values in this indicator are generally better, as they show that an organization has substantial equity in
its assets. Low or negative values indicate an organization has higher liabilities and is generally more
leveraged and thus more vulnerable to shocks. However, low values also indicate that an organization may
be investing more of its equity for growth.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x2[ c(15,300,600) ] <- 0
dat <- data.frame( x1,x2 )
d <- get_er( df=dat, debt='x1', net.assets='x2' )
head( d )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_er( df=dat, debt='x1', net.assets='x2', winsorize=0.95 )

}
