% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-ggr.R
\name{get_ggr}
\alias{get_ggr}
\title{Government Grants Ratio}
\usage{
get_ggr(df, ggc, total.revenue, winsorize = 0.98)
}
\arguments{
\item{df}{A dataframe containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{ggc}{A character string indicating the column name for government grant contributions (On 990: Part VIII, Line 1(E); On EZ: Not Available).}

\item{total.revenue}{A character string indicating the column name for total revenue (On 990: (Part VIII, Line 12A); On EZ: Part I, Line 9).}
}
\value{
The original dataframe appended with the government grants ratio (\code{ggr}),
a winsorized version (\code{ggr.w}), a standardized z-score version (\code{ggr.z}),
and a percentile version (\code{ggr.p}).
}
\description{
Calculate the government grants ratio and append it to the dataframe.
}
\details{
The government grants ratio measures the percentage of an organization’s revenues that come from
government grants and contributions. High values in this ratio indicate that more of an organization's
revenues come from government grants while a lower number indicates that an organization's funds come from
other donation sources (private foundations), earned income, or investment income.
}
\examples{
x1<-rnorm( 1000,100,30 )
x2<-rnorm( 1000,200,30 )
x2[ c(1,10,100) ] <- 0

dat<-data.frame( x1, x2 )

a<-get_ggr( df=dat, ggc='x1', total.revenue='x2', winsorize=0.98)

head( a )

}
