% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-iidr.R
\name{get_iidr}
\alias{get_iidr}
\title{Investment Income Dependence Ratio}
\usage{
get_iidr(
  df,
  invest.income,
  bond.proceeds,
  rent.income,
  other.income,
  total.revenue,
  winsorize = 0.98
)
}
\arguments{
\item{df}{A dataframe containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{invest.income}{A character string indicating the column name for investment income (On 990: Part VIII, Line 3(A); On EZ: Not Available).}

\item{bond.proceeds}{A character string indicating the column name for investment income from tax-exempt bond proceeds (On 990: Part VIII, Line 4(A); On EZ: Not Available).}

\item{rent.income}{A character string indicating the column name for gross rent income (On 990: Part VIII, Line 6(A); On EZ: Not Available).}

\item{other.income}{A character string indicating the column name for gross income from sales of assets other than inventory  (On 990: Part VIII, Line 7(A); On EZ: Not Available).}

\item{total.revenue}{A character string indicating the column name for total revenue (On 990: (Part VIII, Line 12A); On EZ: Part I, Line 9).}
}
\value{
The original dataframe appended with the investment income dependence ratio (\code{iidr}),
a winsorized version (\code{iidr.w}), a standardized z-score version (\code{iidr.z}),
and a percentile version (\code{iidr.p}).
}
\description{
Calculate the investment income dependence ratio and append it to the dataframe.
}
\details{
This metric shows how much of an organization’s total revenues come from investment income
(interest, dividents, gains/losses on sales of securities or other assets). High levels in this metric
indicate an organization is more depenent on investment income (and thus vulnerable to market downturns),
and low values indicate their income comes more from donations or earned income.
}
\examples{
x1<-rnorm( 1000,100,30 )
x2<-rnorm( 1000,200,30 )
x3<-rnorm( 1000,200,30 )
x4<-rnorm( 1000,100,30 )
x5<-rnorm( 1000,200,30 )
x5[ c(1,10,100) ] <- 0

dat<-data.frame( x1, x2, x3, x4, x5 )

a<-get_iidr( df=dat, invest.income='x1', bond.proceeds='x2', rent.income='x3', other.income='x4', total.revenue='x5', winsorize=0.98)

head( a )

}
