% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-lar.R
\name{get_lar}
\alias{get_lar}
\title{Self Sufficiency Ratio}
\usage{
get_lar(df, land, assets, winsorize = 0.98)
}
\arguments{
\item{land}{Land, buildings, and equipment (Part X, line 10b; On EZ: On EZ: Not Available).}

\item{assets}{Total assets (On 990: Part X, Line 16B; On EZ: Part II, line 25B).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
The original dataframe appended with the self sufficiency ratio (\code{lar}),
a winsorized version (\code{lar.w}), a standardized z-score version (\code{lar.z}),
and a percentile version (\code{lar.p}).
}
\description{
Calculate the self sufficiency ratio and append it to the dataframe.
}
\details{
This metric shows the proportion of an organization’s total assets that are made up by land and
buildings. It shows how diversified their asset portfolio is and somewhat how specialized their line of
work is within housing or building development. Higher values mean that an organization has more of its
asset portfolio in land and buildings, and lower values mean that development or land and building holdings
represent a small share of what the organization’s activities and finances are tied up in.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x2[ c(15,300,600) ] <- 0
dat <- data.frame( x1,x2 )
d <- get_lar( df=dat, land='x1', assets='x2' )
head( d )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_lar( df=dat, land='x1', assets='x2', winsorize=0.95 )

}
