% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-lar.R
\name{get_lar}
\alias{get_lar}
\title{Lands to Assets Ratio}
\usage{
get_lar( df, land = "F9_10_ASSET_LAND_BLDG_DEPREC", assets = "F9_10_ASSET_TOT_EOY", winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{land}{Land, buildings, and equipment (Part X, line 10B; On EZ: On EZ: Not Available).}

\item{assets}{Total assets (On 990: Part X, Line 16B; On EZ: Part II, line 25B).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the lands to assets ratio (\code{lar}),
a winsorized version (\code{lar.w}), a standardized z-score version (\code{lar.z}),
and a percentile version (\code{lar.p}).
}
\description{
Calculate the lands to assets ratio and append it to the dataframe.
}
\details{
This metric shows the proportion of an organization’s total assets that are made up by land and
buildings. It shows how diversified their asset portfolio is and somewhat how specialized their line of
work is within housing or building development. Higher values mean that an organization has more of its
asset portfolio in land and buildings, and lower values mean that development or land and building holdings
represent a small share of what the organization’s activities and finances are tied up in. Note: computation of
this metric is available to full 990 filers only.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000, 100, 30 )
x2 <- rnorm( 1000, 200, 30 )
x2[ c( 15, 300, 600 ) ] <- 0

dat <- data.frame( x1, x2 )

# specify own column names
d <- get_lar( df = dat, land = "x1", assets = "x2" )

head( d )

# run with default column names
x1[ seq( from = 1, to = 1000, 50 ) ] <- NA
x2[ seq( from = 1, to = 1000, 71 ) ] <- NA

dat_01 <- data.frame( x1, x2 )

colnames( dat_01 ) <- c( "F9_10_ASSET_LAND_BLDG_DEPREC", "F9_10_ASSET_TOT_EOY" )

d <- get_lar( dat_01 )

# coerce one column to factor
dat_01$F9_10_ASSET_LAND_BLDG_DEPREC <- as.factor( dat_01$F9_10_ASSET_LAND_BLDG_DEPREC )

d <- get_lar( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_lar( df = dat, land = "x1", assets = "x2", winsorize = 0.95 )

d <- get_lar( dat_01, winsorize = 0.95 )


# using 990 data
data( part010810 )
d <- get_lar( df = part010810 )

# now coerce one of the variables to numeric
part010810$F9_10_ASSET_LAND_BLDG_DEPREC <- as.character( part010810$F9_10_ASSET_LAND_BLDG_DEPREC )

d <- get_lar( df = part010810 )

\dontrun{
## Errors ##

# numerator not specified
d <- get_lar( df = dat, land = NULL, assets = "x2" )

# denominator not specified
d <- get_lar( df = dat, land = "x1", assets = NULL )

# neither numerator nor denominator specified
d <- get_lar( df = dat, land = NULL, assets = NULL )

# column names vector not of correct length
d <- get_lar( df = dat, land = c("e","b","c"), assets = "e")

# column names vector not of correct length
d <- get_lar( df = dat, land = "e", assets = c( "e", "b", "c") )
}
}
