% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-moch.R
\name{get_moch}
\alias{get_moch}
\title{Months of Operating Cash on Hand}
\usage{
get_moch(
  df,
  f.cash = NULL,
  f.si = NULL,
  f.pr = NULL,
  f.ar = NULL,
  f.tfe = NULL,
  f.dda = NULL,
  ez.csi = NULL,
  ez.toe = NULL,
  winsorize = 0.98
)
}
\arguments{
\item{f.cash}{Cash, EOY (On 990: Part X, line 1B; On EZ:Part I, line 22 (cash and short-term investments only)).}

\item{f.si}{Short-term investments, EOY (On 990: Part X, line 2B; On EZ: Not Available).}

\item{f.pr}{Pledges and grant receivables, EOY (On 990: Part X, line 3B; On EZ: Not Available).}

\item{f.ar}{Accounts receivables, EOY (On 990: Part X, line 4B; On EZ: Not Available).}

\item{f.tfe}{Total functional expenses (On 990: Part IX, line 25A; On EZ: Not Available).}

\item{f.dda}{Depreciation, depletion, and amortization (On 990: Part IX, line 22A; On EZ: Not Available).}

\item{ez.csi}{Cash, savings, and investment, EOY (On EZ:Part II, line 22B (cash and short-term investments only)).}

\item{ez.toe}{Total operating expenses, EOY (On EZ: Part I, line 17 (operating expenses only)).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
The original dataframe appended with the months of operating cash on hand (\code{moch}),
a winsorized version (\code{moch.w}), a standardized z-score version (\code{moch.z}),
and a percentile version (\code{moch.p}).
}
\description{
Calculate the months of operating cash on hand and append it to the dataframe.
}
\details{
Months of cash on hand is the number of months that an organization can continue to pay its
operating expenses, given the amount of cash available. This metric is good to review at startup of
organization, during periods of low revenues, and prior to undertaking new major activity.The metric is
limited though in the following ways. First, it is based on an average monthly cash outflow, which is not
really the case. Instead, cash tends to be spent in a lumpy manner, such as when rent or payroll are paid.
Also, management tends to take drastic action to reduce expenses as cash reserves decline, so that the
actual months of operation tend to be longer than indicated by this ratio. Thus, it is better to use a
detailed cash flow analysis to determine the precise duration of the available cash, with regular updates.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x3 <- rnorm( 1000,200,30 )
x4 <- rnorm( 1000,200,30 )
x5 <- rnorm( 1000,200,30 )
x6 <- rnorm( 1000,200,30 )
dat <- data.frame( x1, x2, x3, x4, x5, x6 )
a <- get_moch( df=dat, f.cash='x1', f.si='x2',
             f.pr='x3', f.ar='x4', f.tfe='x5', f.dda='x6' )

b <- get_moch( df=dat, f.cash='x1', f.si = 'x2',
              f.pr='x3', f.ar='x4', ez.toe='x5', f.dda='x6' )

x5[ c(1:10) ] <- 0
x6[ c(1:10) ] <- 0
dat <- data.frame( x1, x2, x3, x4, x5, x6 )

c <- get_moch( df=dat, f.cash='x1', f.si='x2',
              f.pr='x3', f.ar='x4', f.tfe='x5', f.dda='x6' )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_moch( df=dat, f.cash='x1', f.si='x2',
              f.pr='x3', f.ar='x4', f.tfe='x5', f.dda='x6', winsorize=0.95 )

}
