% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-efile-data.R
\name{get_ntmaj12}
\alias{get_ntmaj12}
\title{Convert NTEE V2 codes to NTMAJ12 major group labels}
\usage{
get_ntmaj12(nteev2)
}
\arguments{
\item{nteev2}{Character vector of NTEE V2 codes.}
}
\value{
Factor vector with 12 levels (ARTS, EDUCATION, HIGHER ED, etc.).
}
\description{
Convert NTEE V2 codes to NTMAJ12 major group labels
}
\examples{
get_ntmaj12( c("HOS-E05-RG", "EDU-B29-RG") )
}
