% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-or.R
\name{get_or}
\alias{get_or}
\title{Operating Margin}
\usage{
get_or(df, equity.eoy, equity.boy, winsorize = 0.98)
}
\arguments{
\item{df}{A dataframe containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{equity.eoy}{Unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available).}

\item{equity.boy}{Unrestricted net assets, BOY (On 990: Part X, line 27A; On EZ: Not Available).}
}
\value{
The original dataframe appended with the operating margin (\code{or}),
a winsorized version (\code{or.w}), a standardized z-score version (\code{or.z}),
and a percentile version (\code{or.p}).
}
\description{
Calculate the operating margin and append it to the dataframe.
}
\details{
The operating margin measures what percent of an organizations unrestricted net assets (or free
cash) from the beginning of 2019 it had at the end of 2019.
}
\examples{
x1<-rnorm( 1000,100,30 )
x2<-rnorm( 1000,200,30 )
x2[ c(1,10,100) ] <- 0

dat<-data.frame( x1, x2 )

a<-get_or( df=dat, equity.eoy='x1', equity.boy='x2', winsorize=0.98)

head( a )

}
