% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-or.R
\name{get_or}
\alias{get_or}
\title{Operating Margin}
\usage{
get_or( df, 
equity.eoy = "F9_10_NAFB_UNRESTRICT_EOY", 
equity.boy = "F9_10_NAFB_UNRESTRICT_BOY", 
winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{equity.eoy}{Unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available).}

\item{equity.boy}{Unrestricted net assets, BOY (On 990: Part X, line 27A; On EZ: Not Available).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the operating margin (\code{or}),
a winsorized version (\code{or.w}), a standardized z-score version (\code{or.z}),
and a percentile version (\code{or.p}).
}
\description{
Calculate the operating margin and append it to the dataframe.
}
\details{
The operating margin measures what percent of an organizations unrestricted net assets (or free
cash) from the beginning of 2019 it had at the end of 2019. Note: computation of this metric is available
to only 990 filers and not for 990-EZ filers. The default inputs use column names for variables available
only to 990 filers.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000, 100, 30 )
x2 <- rnorm( 1000, 200, 30 )
x2[ c( 15, 300, 600 ) ] <- 0

dat <- data.frame( x1, x2 )

# specify own column names
d <- get_or( df = dat, equity.eoy = "x1", equity.boy = "x2" )

head( d )

# run with default column names
x1[ seq( from = 1, to = 1000, 50 ) ] <- NA
x2[ seq( from = 1, to = 1000, 71 ) ] <- NA

dat_01 <- data.frame( x1, x2 )

colnames( dat_01 ) <- c( "F9_10_NAFB_UNRESTRICT_EOY", "F9_10_NAFB_UNRESTRICT_BOY" )

d <- get_or( dat_01 )

# coerce one column to factor
dat_01$F9_10_NAFB_UNRESTRICT_EOY <- as.factor( dat_01$F9_10_NAFB_UNRESTRICT_EOY )

d <- get_or( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_or( df = dat, equity.eoy = "x1", equity.boy = "x2", winsorize = 0.95 )

d <- get_or( dat_01, winsorize = 0.95 )


# using 990 data
data( part010810 )
d <- get_or( df = part010810 )

# now coerce one of the variables to numeric
part010810$F9_10_NAFB_UNRESTRICT_EOY <- as.character( part010810$F9_10_NAFB_UNRESTRICT_EOY )

d <- get_or( df = part010810 )

\dontrun{
## Errors ##

# numerator not specified
d <- get_or( df = dat, equity.eoy = NULL, equity.boy = "x2" )

# denominator not specified
d <- get_or( df = dat, equity.eoy = "x1", equity.boy = NULL )

# neither numerator nor denominator specified
d <- get_or( df = dat, equity.eoy = NULL, equity.boy = NULL )

# column names vector not of correct length
d <- get_or( df = dat, equity.eoy = c("e","b","c"), equity.boy = "e")

# column names vector not of correct length
d <- get_or( df = dat, equity.eoy = "e", equity.boy = c( "e", "b", "c") )
}

}
