% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-efile-data.R
\name{get_org_type}
\alias{get_org_type}
\title{Classify nonprofit organization type from NTEE code}
\usage{
get_org_type(x)
}
\arguments{
\item{x}{Character vector of NTEE codes.}
}
\value{
Character vector of organization type codes:
AA (alliance), MT (management/technical), PA (planning/advocacy),
RP (research), MS (single-org support), MM (multi-org support),
NS (general nonprofit support), RG (regular program organization).
}
\description{
Determines the functional type of the organization (regular, advocacy,
research, funding, etc.) based on digits 2-3 of the NTEE code.
}
\examples{
get_org_type( c("B01", "B02", "B05", "B11", "B29") )
}
