% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-per.R
\name{get_per}
\alias{get_per}
\title{Program Efficiency Ratio}
\usage{
get_per(df, pse, total.expense, winsorize = 0.98)
}
\arguments{
\item{df}{A dataframe containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{pse}{A character string indicating the column name for program service expenses (On 990: Part 9, Line 25B; On EZ: Pt II, Line 25B).}

\item{total.expense}{A character string indicating the column name for total expenses (On 990: Part 1, Line 18(B); On EZ: Part 1, Line 17).}
}
\value{
The original dataframe appended with the program efficiency ratio (\code{per}),
a winsorized version (\code{per.w}), a standardized z-score version (\code{per.z}),
and a percentile version (\code{per.p}).
}
\description{
Calculate the program efficiency ratio and append it to the dataframe.
}
\details{
The program efficiency ratio measures the percentage of expenses that a nonprofit organization is
spending on its core mission. High values in this ratio indicate that more of an organization's expenses
are going towards its core mission or program services while a lower number indicates that an organization's
expenses are comprised of other things like management or fundraising. Charity Navigator generally gives
the highest rankings to those organizations whose ratio of program expenses is 85\% or higher of their total
expenses. Other agencies, such as the Better Business Bureau’s Wise Giving Alliance, recommend a ratio of
65\% or higher.
}
\examples{
x1<-rnorm( 1000,100,30 )
x2<-rnorm( 1000,200,30 )
x2[ c(1,10,100) ] <- 0

dat<-data.frame( x1, x2 )

a<-get_per( df=dat, pse='x1', total.expense='x2' )

head( a )

}
