% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-podpm.R
\name{get_podpm}
\alias{get_podpm}
\title{Post-Depreciation Profitability Margin}
\usage{
get_podpm(df, expenses, revenue, winsorize = 0.98)
}
\arguments{
\item{expenses}{A character string indicating the column name for total functional expenses (On 990: Part IX, line 25A; On EZ: Not Available).}

\item{revenue}{A character string indicating the column name for total revenue, (On 990: Part VIII, line 12A; On EZ: Part I, line 9).}
}
\value{
The original dataframe appended with the post-depreciation profitability margin (\code{podpm}),
a winsorized version (\code{podpm.w}), a standardized z-score version (\code{podpm.z}),
and a percentile version (\code{podpm.p}).
}
\description{
Calculate the post-depreciation profitability margin and append it to the dataframe.
}
\details{
Post-depreciation profit is an income measure used to determine profit after incorporating
non-cash expenses on a balance sheet. Post-depreciation profit is calculated because it provides a
picture of an organization’s true available profits net of depreciation expenses. Non-expense items lower
an organization’s reported earnings, so a post-depreciation profit would show a lower profit in comparison
to profits calculated prior to depreciation expenses.High values in this metric are generally desirable
since they indicate that an organization is not losing a lot of its revenues to expenses. Values close to
zero are normal, and negative numbers indicate the organization is functioning at a deficit.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x3 <- rnorm( 1000,200,30 )
dat <- data.frame( x1, x2, x3 )
a <- get_podpm( df=dat, revenue='x1', depreciation='x2' )

}
