% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-predpm.R
\name{get_predpm}
\alias{get_predpm}
\title{Pre-Depreciation Profitability Margin}
\usage{
get_predpm(df, expenses, depreciation, revenue, winsorize = 0.98)
}
\arguments{
\item{expenses}{A character string indicating the column name for total functional expenses (On 990: Part IX, line 25A; On EZ: Not Available).}

\item{depreciation}{A character string indicating the column name for depreciation expenses (On 990: Part X, line 2B; On EZ:Not available).}

\item{revenue}{A character string indicating the column name for total revenue (On 990: Part VIII, line 12A; On EZ: Part I, line 9).}
}
\value{
The original dataframe appended with the pre-depreciation profitability margin (\code{predpm}),
a winsorized version (\code{predpm.w}), a standardized z-score version (\code{predpm.z}),
and a percentile version (\code{predpm.p}).
}
\description{
Calculate the pre-depreciation profitability margin and append it to the dataframe.
}
\details{
Pre-depreciation profit is an income measure used to determine profit before incorporating
non-cash expenses on a balance sheet. Pre-depreciation profit is calculated because it provides a cleaner
number that can help determine a organization’s ability to service debt. Much like free cash flow,
pre-depreciation profit is a measure of a organization’s actual cash flow. Non-expense items lower an
organization’s reported earnings, so a pre-depreciation profit would show a higher profit in comparison to
profits calculated after depreciation. High values in this metric are generally desirable since they
indicate that an organization is not losing a lot of its revenues to expenses, though the amount of
expense exempted from this metric due to depreciation (which for community development corporations can
represent a large portion of their expenses), makes it less of an indicator of true profitability and more
an indicator of an organization’s cash flow. Values close to zero are normal, and negative numbers
indicate the organization is functioning at a deficit.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x3 <- rnorm( 1000,200,30 )
dat <- data.frame( x1, x2, x3 )
a <- get_predpm( df=dat, revenue='x1', depreciation='x2',
             expenses='x3' )

}
