% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-predpm.R
\name{get_predpm}
\alias{get_predpm}
\title{Pre-Depreciation Profitability Margin}
\usage{
get_predpm( df, 
expenses = "F9_09_EXP_TOT_TOT",
revenue = "F9_08_REV_TOT_TOT", 
depreciation = "F9_09_EXP_DEPREC_TOT", 
numerator = NULL,
denominator = NULL,
winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{expenses}{A character string indicating the column name for total functional expenses (On 990: Part IX, line 25A; On EZ: Not Available).}

\item{revenue}{A character string indicating the column name for total revenue (On 990: Part VIII, line 12A; On EZ: Part I, line 9).}

\item{depreciation}{A character string indicating the column name for depreciation expenses (On 990: Part IX, line 22A; On EZ:Not available).}

\item{numerator}{A character string indicating the user-supplied column name for a pre-aggregated variable for the numerator. Do not combine with numerator column component arguments (\code{revenue}, \code{expenses},\code{depreciation}).}

\item{denominator}{A character string indicating the user-supplied column name for a pre-aggregated variable for the denominator. Do not combine with denominator column component arguments (\code{revenue}).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the pre-depreciation profitability margin (\code{predpm}),
a winsorized version (\code{predpm.w}), a standardized z-score version (\code{predpm.z}),
and a percentile version (\code{predpm.p}).
}
\description{
Calculate the pre-depreciation profitability margin and append it to the dataframe.
}
\details{
Pre-depreciation profit is an income measure used to determine profit before incorporating
non-cash expenses on a balance sheet. Pre-depreciation profit is calculated because it provides a cleaner
number that can help determine a organization’s ability to service debt. Much like free cash flow,
pre-depreciation profit is a measure of a organization’s actual cash flow. Non-expense items lower an
organization’s reported earnings, so a pre-depreciation profit would show a higher profit in comparison to
profits calculated after depreciation. High values in this metric are generally desirable since they
indicate that an organization is not losing a lot of its revenues to expenses, though the amount of
expense exempted from this metric due to depreciation (which for community development corporations can
represent a large portion of their expenses), makes it less of an indicator of true profitability and more
an indicator of an organization’s cash flow. Values close to zero are normal, and negative numbers
indicate the organization is functioning at a deficit.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000, 100, 30 )
x2 <- rnorm( 1000, 200, 30 )
x3 <- rnorm( 1000, 200, 30 )
x3[ c( 15, 300, 600 ) ] <- 0

dat <- data.frame( x1, x2, x3 )

# specify own column names
d <- get_predpm( df = dat, expenses = "x1", depreciation = "x2", revenue = "x3" )


head( d ) 

# run with default column names
x1[ seq( from = 1, to = 1000, 50 ) ] <- NA
x2[ seq( from = 1, to = 1000, 71 ) ] <- NA

dat_01 <- data.frame( x1, x2, x3 )

colnames( dat_01 ) <- c( "F9_09_EXP_TOT_TOT", "F9_09_EXP_DEPREC_TOT", "F9_08_REV_TOT_TOT" )

d <- get_predpm( dat_01 )

# coerce one column to factor
dat_01$F9_09_EXP_TOT_TOT <- as.factor( dat_01$F9_09_EXP_TOT_TOT )

d <- get_predpm( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_predpm( df = dat, expenses = "x1", depreciation = "x2", revenue = "x3", winsorize = 0.95 )

d <- get_predpm( dat_01, winsorize = 0.95 )


# using 990 data
data( part010810 )
d <- get_predpm( df = part010810 )

# now coerce one of the variables to numeric
part010810$F9_09_EXP_TOT_TOT <- as.character( part010810$F9_09_EXP_TOT_TOT )

d <- get_predpm( df = part010810 )

\dontrun{
## Errors ##

# numerator not specified
d <- get_predpm( df = dat, expenses = NULL, depreciation = NULL, revenue = "x3" )

# denominator not specified
d <- get_predpm( df = dat, expenses = "x1", depreciation = "x2", revenue = NULL )

# neither numerator nor denominator specified
d <- get_predpm( df = dat, expenses = NULL, depreciation = NULL, revenue = NULL )

# column names vector not of correct length
d <- get_predpm( df = dat, expenses = c("e","b","c"), depreciation = "x2", revenue = "e")

# column names vector not of correct length
d <- get_predpm( df = dat, expenses = "e", revenue = c( "e", "b", "c"), depreciation = "x2" )
}
}
