% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-qr.R
\name{get_qr}
\alias{get_qr}
\title{Quick Ratio}
\usage{
get_qr(df, cash, si, pr, ar, ap, gp, winsorize = 0.98)
}
\arguments{
\item{cash}{A character string indicating the column name for cash, EOY (On 990: Part X, line 1B; On EZ: Not Available).}

\item{si}{A character string indicating the column name for short-term investments, EOY (On 990: Part X, line 2B;On EZ: Not Available).}

\item{pr}{A character string indicating the column name for pledges and grant receivables, EOY On 990: Part X, line 3B; On EZ: Not Available).}

\item{ar}{A character string indicating the column name for accounts receivables, EOY (On 990: Part X, line 4B; On EZ: Not Available).}

\item{ap}{A character string indicating the column name for accounts payable, EOY (On 990: Part X, line 17B; On EZ: Not Available).}

\item{gp}{A character string indicating the column name for grants payable, EOY (On 990: Part X, line 18B; On EZ: Not Available).

#' @return The original dataframe appended with the quick ratio (\code{qr}),
a winsorized version (\code{qr.w}), a standardized z-score version (\code{qr.z}),
and a percentile version (\code{qr.p}).}
}
\description{
Calculate the quick ratio and append it to the dataframe.
}
\details{
This metric indicates how well an organization can cover its liabilities with its readily
available cash. When an organization has a quick ratio of 1, its quick assets are equal to its current
liabilities. This also indicates that the organization can pay off its current debts without selling its
long-term assets. If an organization has a quick ratio higher than 1, this means that it owns more quick
assets than current liabilities.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x3 <- rnorm( 1000,200,30 )
x4 <- rnorm( 1000,100,30 )
x5 <- rnorm( 1000,200,30 )
x6 <- rnorm( 1000,200,30 )
dat <- data.frame( x1, x2, x3, x4, x5, x6 )

a <- get_qr( df=dat, cash='x1', si='x2', pr='x3', ar='x4', ap='x5', gp='x6', winsorize=0.98 )

}
