% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-ssr.R
\name{get_ssr}
\alias{get_ssr}
\title{Self Sufficiency Ratio}
\usage{
get_ssr(df, debt, net.assets, winsorize = 0.98)
}
\arguments{
\item{debt}{Total debt, EOY (On 990: Part X, line 17B; On EZ: Not Available).}

\item{net.assets}{Unrestricted net assets, EOY (On 990: Part X, line 27B; On EZ: Not Available).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
The original dataframe appended with the self sufficiency ratio (\code{ssr}),
a winsorized version (\code{ssr.w}), a standardized z-score version (\code{ssr.z}),
and a percentile version (\code{ssr.p}).
}
\description{
Calculate the self sufficiency ratio and append it to the dataframe.
}
\details{
The Self Sufficiency Ratio measures the proportion of operating expenses that are covered by earned income.
This metric is a good measure of how long an organization can survive without grants. Higher values
mean it is more self-sufficient, meaning it could cover its costs longer without collecting any grants,
rents, royalties, or sales of inventory. This ratio is primarily useful for organizations that have
earned revenue through developers’ fees, management fees, memberships, or tuition. Higher values mean
organizations have more autonomy and flexibility. They generally improve over time as an organization
grows. In the early stages, these ratios tend to be lower but the goal is to make them as high as possible.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x2[ c(15,300,600) ] <- 0
dat <- data.frame( x1,x2 )
d <- get_ssr( df=dat, debt='x1', net.assets='x2' )
head( d )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_ssr( df=dat, debt='x1', net.assets='x2', winsorize=0.95 )

}
