% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal-health-package-stdr.R
\name{get_stdr}
\alias{get_stdr}
\title{Short Term Debt Ratio}
\usage{
get_stdr(df, ap, gp, net.assets, winsorize = 0.98)
}
\arguments{
\item{ap}{A character string indicating the column name for accounts payable, EOY (On 990: Part X, line 17B; On EZ: Not Available).}

\item{gp}{A character string indicating the column name for grants payable, EOY (On 990: Part X, line 18B; On EZ: Not Available).}

\item{net.assets}{A character string indicating the column name for net assets, EOY (On 990: Part X, Line 33B; On EZ: Part I, Line 21).}
}
\value{
The original dataframe appended with the short term debt ratio (\code{stdr}),
a winsorized version (\code{stdr.w}), a standardized z-score version (\code{stdr.z}),
and a percentile version (\code{stdr.p}).
}
\description{
Calculate the short term debt ratio and append it to the dataframe.
}
\details{
This metric indicates how well an organization can cover its immediate liabilities with its equity.
Low values are better as they indicate that an organization has lower levels of leverage and more flexibility
to dispense its assets as needed or as opportunities arise. When an organization has a Short Term Debt Ratio
of 1, its liabilities are fully equal to the assets it owns. If an organization has a ratio higher than 1, it
indicates that it is over leveraged. A value close to zero indicates low levels of liability, and a negative
value indicates that an organization has overpaid on their liabilities.
}
\examples{
x1 <- rnorm( 1000,100,30 )
x2 <- rnorm( 1000,200,30 )
x3 <- rnorm( 1000,200,30 )
x3[ c(1,10,100)] <- 0
dat <- data.frame( x1, x2, x3 )

a <- get_stdr( df=dat, ap='x1', gp='x2', net.assets='x3', winsorize=0.98 )

}
