% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stdr.R
\name{get_stdr}
\alias{get_stdr}
\title{Short Term Debt Ratio}
\usage{
get_stdr( df, accounts.payable = "F9_10_LIAB_ACC_PAYABLE_EOY", grants.payable = "F9_10_LIAB_GRANT_PAYABLE_EOY", net.assets = "F9_10_NAFB_TOT_EOY", numerator = NULL, denominator = NULL, winsorize = 0.98 )
}
\arguments{
\item{df}{A \code{data.frame} containing the required field for computing the metric. The metric will be appended to this dataset.}

\item{accounts.payable}{A character string indicating the column name for accounts payable, EOY (On 990: Part X, line 17B; On EZ: Not Available).}

\item{grants.payable}{A character string indicating the column name for grants payable, EOY (On 990: Part X, line 18B; On EZ: Not Available).}

\item{net.assets}{A character string indicating the column name for net assets, EOY (On 990: Part X, Line 33B; On EZ: Part I, Line 21).}

\item{numerator}{A character string indicating the user-supplied column name for a pre-aggregated variable for the numerator. Do not combine with numerator column component arguments (\code{accounts.payable}, \code{grants.payable}).}

\item{denominator}{A character string indicating the user-supplied column name for a pre-aggregated variable for the denominator. Do not combine with denominator column component arguments (\code{net.assets}).}

\item{winsorize}{The winsorization value (between 0 and 1), defaults to 0.98 which winsorizes at 99th and 1st percentile values.}
}
\value{
Object of class \code{data.frame}: the original dataframe appended with the short term debt ratio (\code{stdr}),
a winsorized version (\code{stdr.w}), a standardized z-score version (\code{stdr.z}),
and a percentile version (\code{stdr.p}).
}
\description{
Calculate the short term debt ratio and append it to the dataframe.
}
\details{
This metric indicates how well an organization can cover its immediate liabilities with its equity.
Low values are better as they indicate that an organization has lower levels of leverage and more flexibility
to dispense its assets as needed or as opportunities arise. When an organization has a Short Term Debt Ratio
of 1, its liabilities are fully equal to the assets it owns. If an organization has a ratio higher than 1, it
indicates that it is over leveraged. A value close to zero indicates low levels of liability, and a negative
value indicates that an organization has overpaid on their liabilities. Note: computation of this metric is available
to only 990 filers and not for 990-EZ filers. The default inputs use column names for variables available only to 990
filers.
}
\examples{
library( fiscal )
x1 <- rnorm( 1000, 100, 30 )
x2 <- rnorm( 1000, 200, 30 )
x3 <- rnorm( 1000, 200, 30 )
x3[ c( 15, 300, 600 ) ] <- 0

dat <- data.frame( x1, x2, x3 )

# specify own column names
d <- get_stdr( df = dat, accounts.payable = "x1", grants.payable = "x2", net.assets = "x3" )


head( d ) 

# run with default column names
x1[ seq( from = 1, to = 1000, 50 ) ] <- NA
x2[ seq( from = 1, to = 1000, 71 ) ] <- NA

dat_01 <- data.frame( x1, x2, x3 )

colnames( dat_01 ) <- c( "F9_10_LIAB_ACC_PAYABLE_EOY", "F9_10_LIAB_GRANT_PAYABLE_EOY", "F9_10_NAFB_TOT_EOY" )

d <- get_stdr( dat_01 )

# coerce one column to factor
dat_01$F9_10_LIAB_GRANT_PAYABLE_EOY <- as.factor( dat_01$F9_10_LIAB_GRANT_PAYABLE_EOY )

d <- get_stdr( dat_01 )

# winsorize at 0.025 and 0.975 percentiles instead of 0.01 and 0.99
d <- get_stdr( df = dat, accounts.payable = "x1", grants.payable = "x2", net.assets = "x3", winsorize = 0.95 )

d <- get_stdr( dat_01, winsorize = 0.95 )


# using 990 data
data( part010810 )
d <- get_stdr( df = part010810 )

# now coerce one of the variables to numeric
part010810$F9_09_EXP_TOT_TOT <- as.character( part010810$F9_09_EXP_TOT_TOT )

d <- get_stdr( df = part010810 )

\dontrun{
## Errors ##

# numerator not specified
d <- get_stdr( df = dat, accounts.payable = NULL, grants.payable = NULL, net.assets = "x3" )

# denominator not specified
d <- get_stdr( df = dat, accounts.payable = "x1", grants.payable = "x2", net.assets = NULL )

# neither numerator nor denominator specified
d <- get_stdr( df = dat, accounts.payable = NULL, grants.payable = NULL, net.assets = NULL )

# column names vector not of correct length
d <- get_stdr( df = dat, accounts.payable = c("e","b","c"), grants.payable = "x2", net.assets = "e")

# column names vector not of correct length
d <- get_stdr( df = dat, accounts.payable = "e", net.assets = c( "e", "b", "c"), grants.payable = "x2" )
}
}
