% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{part010810}
\alias{part010810}
\title{NCCS 990/990-EZ Data Excerpt}
\format{
A data frame with 1000 rows and 329 variables:
\describe{
Data Dictionary: https://nonprofit-open-data-collective.github.io/irs990efile/data-dictionary/data-dictionary.html}
}
\source{
\url{https://github.com/lecy/fiscal-health/tree/main/EFILE}
}
\usage{
part010810
}
\description{
A dataset containing data from 990 and 990-EZ filings, 2014-2021.
}
\examples{
 data( part010810 )
}
\keyword{datasets}
