% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-x.R
\name{rank_normal_stable}
\alias{rank_normal_stable}
\title{Rank-based inverse normal transform (Blom approximation)}
\usage{
rank_normal_stable(x)
}
\description{
Rank-based inverse normal transform (Blom approximation)
}
\keyword{internal}
